SBZ3 Fixes (Mercury/RetroKoH)*
The button that appears right at level start appears glitched for a moment while the graphics load in (Mercury).
If Sonic moves rightward very quickly into the first slope in the level, he can pass through it. I altered collision data and added a block to resolve this (RetroKoH/Mercury).
S3K Object Manager seems to have fixed issues related to the giant platform and the tiny platform that was missing.