Added SRAM Support and Progress Saving (Kilo/RetroKoH)
PSB object now doubles as a menu if the mod is enabled. Progress is saved to SRAM at the end of each act, and loaded again if you choose CONTINUE at the title menu. Progress is reset if you choose NEW GAME. Nothing will be saved if you activate Level Select.