Adjusted Invisibarriers in SBZ2
There was one barrier that I shifted down $20 pixels (next to a lone Sliding Door (Obj6B)) due to what seemed to be a minor collision issue.

If I shift the door one pixel to the left, it no longer appears behind the wall, but there is a bug where Sonic stands on it, within the wall.

Something needs to be done with collision to fix this.